third test

+++ TP2/tp2/ex1_image_processing.cu
@@ -59,7 +59,7 @@ int main(int argc, char const *argv[])
     dim3 grid((M + block.x - 1) / block.x, (N + block.y - 1) / block.y);
 
     // Call the kernel function to process the image
-    process<<<grid, block>>>(N, M, C, (int)pitch / sizeof(float), d_img);
+    process<<<grid, block>>>(N, M, C, pitch / sizeof(float), d_img);
 
     // Copy the modified image back to the host
     cudaMemcpy2D(img, C * N * sizeof(float), d_img, pitch, C * N * sizeof(float), M, cudaMemcpyDeviceToHost);
